refact: captureSerivce 책임 분리
